update user's token check on update delete create posts' routes

+++ backend/src/posts/posts.service.ts
@@ -5,10 +5,14 @@ import {EntityManager, Repository} from "typeorm";
 import {Post} from "./entities/post.entity";
 import {InjectRepository} from "@nestjs/typeorm";
 import {User} from "../users/entities/user.entity";
+import {SecurityService} from "../users/user-security.service";
+import {Request, Response} from "express";
+import {AuthToken} from "../users/entities/auth.token";
 
 @Injectable()
 export class PostsService {
   constructor(
+      private readonly securityService: SecurityService,
       @InjectRepository(Post)
       private readonly postRepository: Repository<Post>,
       @InjectRepository(User)
@@ -30,25 +34,51 @@ export class PostsService {
     return await this.postRepository.find({ where: { user: authorUser } });
   }
 
-  async create(createPostDto: CreatePostDto): Promise<void> {
-    const currentUser = await this.userRepository.findOne({ where: { username: createPostDto.username } });
+  async create(createPostDto: CreatePostDto, req : Request, res : Response): Promise<void> {
 
-    if(!currentUser) {
-      throw new NotFoundException("User not found");
+    let currentUser;
+
+    try{
+      currentUser = await this.securityService.getUserByCookie(req, res);
     }
+    catch(err){
+      res.status(401).send('Invalid authentication');
+      return;
+    }
+
 
     const post = this.entityManager.create(Post, {
       ...createPostDto,
-      user: currentUser,
+      user : currentUser
     });
 
     try {
       await this.entityManager.save(post);
+      res.status(201).send('Successfully created post');
     } catch (error) {
       throw new NotFoundException('Failed to save the post');
     }
   }
-  async update(id: number, updatePostDto: UpdatePostDto): Promise<void> {
+  async update(id: number, updatePostDto: UpdatePostDto, req : Request, res : Response): Promise<void> {
+    let currentUser;
+
+    try {
+      currentUser = await this.securityService.getUserByCookie(req, res);
+    }
+    catch (err){
+      res.status(401).send('Invalid authentication');
+      return;
+    }
+
+    const post = await this.postRepository.findOne({where : {id : id}});
+    if(!post){
+      throw new NotFoundException('Post not found');
+    }
+
+    if(post.user !== currentUser){
+      throw new BadRequestException('You cant change this post');
+    }
+
     const result = await this.entityManager
         .createQueryBuilder()
         .update(Post)
@@ -62,9 +92,24 @@ export class PostsService {
     }
   }
 
-  async delete(id: number) {
-    const post = await this.postRepository.findOne({where : {id}});
+  async delete(id: number, req : Request, res : Response): Promise<void> {
+    let currentUser;
+
+    try {
+      currentUser = await this.securityService.getUserByCookie(req, res);
+    }
+    catch (err){
+      res.status(401).send('Invalid authentication');
+      return;
+    }
+
+    const post = await this.postRepository.findOne({where : {id}, relations : ['user']});
     if(!post) return;
+
+    if(post.user !== currentUser){
+      throw new BadRequestException('You cant change this post');
+    }
+
     await this.entityManager.remove(post);
   }
 }
